requirements_backup.txt
streamlit>=1.40.0
opencv-python-headless>=4.8.0,<5.0.0
pillow>=9.0.0,<13.0.0
numpy>=1.20.0,<2.5.0
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
ultralytics>=8.0.0
